+++ src/options.ts
@@ -1,3 +1,3 @@
-import Options from "./components/Options.svelte";
+import Options from "./options/components/Options.svelte";
 
 new Options({ target: document.body });
