+++ src/options/components/Fieldset.svelte
